Additional type migration
