+++ src/command/render/latexmk/latex.ts
@@ -9,10 +9,11 @@ import { basename, join } from "path/mod.ts";
 import { existsSync } from "fs/mod.ts";
 import { error, info } from "log/mod.ts";
 
+import { PdfEngine } from "../../../config/types.ts";
+
 import { dirAndStem } from "../../../core/path.ts";
 import { execProcess, ProcessResult } from "../../../core/process.ts";
 
-import { PdfEngine } from "../../../config/pdf.ts";
 import { PackageManager } from "./pkgmgr.ts";
 import {
   kLatexBodyMessageOptions,
